Decoupled services in hierarchical fashion. Updated tests and other services to support this code organisation.
