+++ src/main/java/rs/teslaris/core/service/impl/DocumentPublicationServiceImpl.java
@@ -2,96 +2,47 @@ package rs.teslaris.core.service.impl;
 
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
-import rs.teslaris.core.converter.JournalPublicationConverter;
 import rs.teslaris.core.dto.document.DocumentDTO;
 import rs.teslaris.core.dto.document.DocumentFileDTO;
-import rs.teslaris.core.dto.document.JournalPublicationDTO;
-import rs.teslaris.core.dto.document.JournalPublicationResponseDTO;
 import rs.teslaris.core.exception.NotFoundException;
 import rs.teslaris.core.model.commontypes.ApproveStatus;
+import rs.teslaris.core.model.commontypes.BaseEntity;
 import rs.teslaris.core.model.document.Document;
-import rs.teslaris.core.model.document.JournalPublication;
 import rs.teslaris.core.repository.document.DocumentRepository;
 import rs.teslaris.core.service.DocumentFileService;
 import rs.teslaris.core.service.DocumentPublicationService;
-import rs.teslaris.core.service.JournalService;
 import rs.teslaris.core.service.MultilingualContentService;
 import rs.teslaris.core.service.PersonContributionService;
 
 @Service
+@Primary
 @RequiredArgsConstructor
 @Transactional
 public class DocumentPublicationServiceImpl implements DocumentPublicationService {
 
-    private final DocumentRepository documentRepository;
+    protected final DocumentRepository documentRepository;
 
-    private final DocumentFileService documentFileService;
+    protected final DocumentFileService documentFileService;
 
-    private final MultilingualContentService multilingualContentService;
-
-    private final JournalService journalService;
+    protected final MultilingualContentService multilingualContentService;
 
     private final PersonContributionService personContributionService;
 
     @Value("${document.approved_by_default}")
-    private Boolean documentApprovedByDefault;
+    protected Boolean documentApprovedByDefault;
 
 
     @Override
     public Document findDocumentById(Integer documentId) {
         return documentRepository.findById(documentId)
             .orElseThrow(() -> new NotFoundException("Document with given id does not exist."));
-    }
-
-    @Override
-    public JournalPublicationResponseDTO readJournalPublicationById(Integer publicationId) {
-        var publication = (JournalPublication) findDocumentById(publicationId);
-        if (!publication.getApproveStatus().equals(ApproveStatus.APPROVED)) {
-            throw new NotFoundException("Document with given id does not exist.");
-        }
-        return JournalPublicationConverter.toDTO(publication);
-    }
-
-    @Override
-    public JournalPublication createJournalPublication(JournalPublicationDTO publicationDTO) {
-        var publication = new JournalPublication();
-
-        setCommonFields(publication, publicationDTO);
-        setJournalPublicationRelatedFields(publication, publicationDTO);
-
-        publication.setApproveStatus(
-            documentApprovedByDefault ? ApproveStatus.APPROVED : ApproveStatus.REQUESTED);
-
-        return documentRepository.save(publication);
-    }
-
-    @Override
-    public void editJournalPublication(Integer publicationId,
-                                       JournalPublicationDTO publicationDTO) {
-        var publicationToUpdate = (JournalPublication) findDocumentById(publicationId);
-
-        clearCommonFields(publicationToUpdate);
-        publicationToUpdate.getUris().clear();
-
-        setCommonFields(publicationToUpdate, publicationDTO);
-        setJournalPublicationRelatedFields(publicationToUpdate, publicationDTO);
-    }
-
-    @Override
-    public void deleteJournalPublication(Integer journalPublicationId) {
-        var publicationToDelete = (JournalPublication) findDocumentById(journalPublicationId);
-
-        publicationToDelete.getProofs()
-            .forEach(proof -> documentFileService.deleteDocumentFile(proof.getServerFilename()));
-
-        publicationToDelete.getFileItems().forEach(
-            fileItem -> documentFileService.deleteDocumentFile(fileItem.getServerFilename()));
-
-        documentRepository.delete(publicationToDelete);
     }
 
     @Override
@@ -136,7 +87,13 @@ public class DocumentPublicationServiceImpl implements DocumentPublicationServic
         documentFileService.deleteDocumentFile(documentFile.getServerFilename());
     }
 
-    private void setCommonFields(Document document, DocumentDTO documentDTO) {
+    @Override
+    public List<Integer> getContributorIds(Integer publicationId) {
+        return findDocumentById(publicationId).getContributors().stream().map(BaseEntity::getId)
+            .filter(Objects::nonNull).collect(Collectors.toList());
+    }
+
+    protected void setCommonFields(Document document, DocumentDTO documentDTO) {
         document.setTitle(
             multilingualContentService.getMultilingualContent(documentDTO.getTitle()));
         document.setSubTitle(
@@ -157,24 +114,11 @@ public class DocumentPublicationServiceImpl implements DocumentPublicationServic
         document.setFileItems(new HashSet<>());
     }
 
-    private void clearCommonFields(Document publication) {
+    protected void clearCommonFields(Document publication) {
         publication.getTitle().clear();
         publication.getSubTitle().clear();
         publication.getDescription().clear();
         publication.getKeywords().clear();
         publication.getContributors().clear();
-    }
-
-    private void setJournalPublicationRelatedFields(JournalPublication publication,
-                                                    JournalPublicationDTO publicationDTO) {
-        publication.setJournalPublicationType(publicationDTO.getJournalPublicationType());
-        publication.setStartPage(publicationDTO.getStartPage());
-        publication.setEndPage(publicationDTO.getEndPage());
-        publication.setNumberOfPages(publicationDTO.getNumberOfPages());
-        publication.setArticleNumber(publicationDTO.getArticleNumber());
-        publication.setVolume(publicationDTO.getVolume());
-        publication.setIssue(publicationDTO.getIssue());
-
-        publication.setJournal(journalService.findJournalById(publicationDTO.getJournalId()));
     }
 }
