+++ src/main/java/rs/teslaris/core/service/impl/JournalPublicationServiceImpl.java
@@ -1,0 +1,95 @@
+package rs.teslaris.core.service.impl;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import rs.teslaris.core.converter.JournalPublicationConverter;
+import rs.teslaris.core.dto.document.JournalPublicationDTO;
+import rs.teslaris.core.dto.document.JournalPublicationResponseDTO;
+import rs.teslaris.core.exception.NotFoundException;
+import rs.teslaris.core.model.commontypes.ApproveStatus;
+import rs.teslaris.core.model.document.JournalPublication;
+import rs.teslaris.core.repository.document.DocumentRepository;
+import rs.teslaris.core.service.DocumentFileService;
+import rs.teslaris.core.service.JournalPublicationService;
+import rs.teslaris.core.service.JournalService;
+import rs.teslaris.core.service.MultilingualContentService;
+import rs.teslaris.core.service.PersonContributionService;
+
+@Service
+public class JournalPublicationServiceImpl extends DocumentPublicationServiceImpl
+    implements JournalPublicationService {
+
+    private final JournalService journalService;
+
+    @Autowired
+    public JournalPublicationServiceImpl(DocumentRepository documentRepository,
+                                         DocumentFileService documentFileService,
+                                         MultilingualContentService multilingualContentService,
+                                         JournalService journalService,
+                                         PersonContributionService personContributionService,
+                                         JournalService journalService1) {
+        super(documentRepository, documentFileService, multilingualContentService,
+            personContributionService);
+        this.journalService = journalService1;
+    }
+
+    @Override
+    public JournalPublicationResponseDTO readJournalPublicationById(Integer publicationId) {
+        var publication = (JournalPublication) findDocumentById(publicationId);
+        if (!publication.getApproveStatus().equals(ApproveStatus.APPROVED)) {
+            throw new NotFoundException("Document with given id does not exist.");
+        }
+        return JournalPublicationConverter.toDTO(publication);
+    }
+
+    @Override
+    public JournalPublication createJournalPublication(JournalPublicationDTO publicationDTO) {
+        var publication = new JournalPublication();
+
+        setCommonFields(publication, publicationDTO);
+        setJournalPublicationRelatedFields(publication, publicationDTO);
+
+        publication.setApproveStatus(
+            documentApprovedByDefault ? ApproveStatus.APPROVED : ApproveStatus.REQUESTED);
+
+        return documentRepository.save(publication);
+    }
+
+    @Override
+    public void editJournalPublication(Integer publicationId,
+                                       JournalPublicationDTO publicationDTO) {
+        var publicationToUpdate = (JournalPublication) findDocumentById(publicationId);
+
+        clearCommonFields(publicationToUpdate);
+        publicationToUpdate.getUris().clear();
+
+        setCommonFields(publicationToUpdate, publicationDTO);
+        setJournalPublicationRelatedFields(publicationToUpdate, publicationDTO);
+    }
+
+    @Override
+    public void deleteJournalPublication(Integer journalPublicationId) {
+        var publicationToDelete = (JournalPublication) findDocumentById(journalPublicationId);
+
+        publicationToDelete.getProofs()
+            .forEach(proof -> documentFileService.deleteDocumentFile(proof.getServerFilename()));
+
+        publicationToDelete.getFileItems().forEach(
+            fileItem -> documentFileService.deleteDocumentFile(fileItem.getServerFilename()));
+
+        documentRepository.delete(publicationToDelete);
+    }
+
+    private void setJournalPublicationRelatedFields(JournalPublication publication,
+                                                    JournalPublicationDTO publicationDTO) {
+        publication.setJournalPublicationType(publicationDTO.getJournalPublicationType());
+        publication.setStartPage(publicationDTO.getStartPage());
+        publication.setEndPage(publicationDTO.getEndPage());
+        publication.setNumberOfPages(publicationDTO.getNumberOfPages());
+        publication.setArticleNumber(publicationDTO.getArticleNumber());
+        publication.setVolume(publicationDTO.getVolume());
+        publication.setIssue(publicationDTO.getIssue());
+
+        publication.setJournal(journalService.findJournalById(publicationDTO.getJournalId()));
+    }
+}
